Test pytables refactor (#28746)
