+++ pandas/util/_test_decorators.py
@@ -27,6 +27,7 @@ from distutils.version import LooseVersion
 import locale
 from typing import Callable, Optional
 
+import numpy as np
 import pytest
 
 from pandas.compat import is_platform_32bit, is_platform_windows
@@ -71,6 +72,21 @@ def safe_import(mod_name, min_version=None):
                 return mod
 
     return False
+
+
+# TODO:
+# remove when gh-24839 is fixed; this affects numpy 1.16
+# and pytables 3.4.4
+tables = safe_import("tables")
+xfail_non_writeable = pytest.mark.xfail(
+    tables
+    and LooseVersion(np.__version__) >= LooseVersion("1.16")
+    and LooseVersion(tables.__version__) < LooseVersion("3.5.1"),
+    reason=(
+        "gh-25511, gh-24839. pytables needs a "
+        "release beyong 3.4.4 to support numpy 1.16x"
+    ),
+)
 
 
 def _skip_if_no_mpl():
